+++ tests/logs/error_fns_without_return_type/.expected
@@ -1,0 +1,14 @@
+error: called function `fn_()` with no return type (at <root>/main__const_global.gpex:6:17)
+  --> info: function has no return type (at <root>/main__const_global.gpex:2:4)
+error: called function `fn_()` with no return type (at <root>/main__fn_param_type.gpex:6:20)
+  --> info: function has no return type (at <root>/main__fn_param_type.gpex:2:4)
+error: called function `fn_()` with no return type (at <root>/main__fn_return_type.gpex:6:17)
+  --> info: function has no return type (at <root>/main__fn_return_type.gpex:2:4)
+error: called function `fn_()` with no return type (at <root>/main__repeat_fn_arg.gpex:6:12)
+  --> info: function has no return type (at <root>/main__repeat_fn_arg.gpex:2:4)
+error: called function `fn_()` with no return type (at <root>/main__statement_assignment_right.gpex:6:24)
+  --> info: function has no return type (at <root>/main__statement_assignment_right.gpex:2:4)
+error: called function `fn_()` with no return type (at <root>/main__statement_return.gpex:6:30)
+  --> info: function has no return type (at <root>/main__statement_return.gpex:2:4)
+error: called function `fn_()` with no return type (at <root>/main__var_global.gpex:6:15)
+  --> info: function has no return type (at <root>/main__var_global.gpex:2:4)
