+++ tests/logs/error_fns_without_return_type/main__$$.gpex
@@ -1,0 +1,6 @@
+var value = 0;
+fn fn_() {
+    value = 1;
+}
+
+{{fn_call_location.code}}
